wip: move authtoken code out of connector
